requirement.txt
asgiref==3.6.0
Django==4.1.4
django-auditlog==2.2.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
